[Feat] #19 MVVM
